Fix bug in RelayCommand and RelayCommand<T>
The 'CommandManager.RequerySuggested' event handler registration was not correct (multiple subscriptions). I just refactored the type on the hurry and did not test it (or properly review it).
You likely don't care. However, for the sake of completeness and for demonstrating reliability I decided to push the fix.

I wish you all the best in regards of health and success for 2024!